+++ terra-rust-api/Changelog.md
@@ -1,8 +1,11 @@
 # Changes
 
 ## 1.0
+### 10.11 - 28-Oct-21
+* tx attribute finder helper function
 ### 1.0.10 - 27-Oct-21
 * tx.timestamp now visible
+* tx.get_events() helper function
 ### 1.0.9 - 26-Oct-21
 * tx.logs can be null
 ### 1.0.8 - 9-Oct-21